Separeted Saved Model iterator from Proto iterator
