+++ src/frontends/tensorflow/src/saved_model.cpp
@@ -5,10 +5,9 @@
 #include <stdlib.h>
 
 #include <fstream>
-#include <map>
 #include <string>
 
-#include "graph_iterator_proto.hpp"
+#include "graph_iterator_saved_model.hpp"
 #include "tensor_bundle.pb.h"
 #include "trackable_object_graph.pb.h"
 
@@ -200,7 +199,7 @@ void GraphIteratorSavedModel::readCMOGraph() {
 
     // Might be need to remove this verification:
     // https://github.com/tensorflow/tensorflow/blob/d90f1947ebcf510b23c238f43c2191e5b3817cb3/tensorflow/cc/experimental/libexport/load.cc#L73
-    FRONT_END_GENERAL_CHECK(result, "CMO: Trackable Object Graph couldn't be read");
+    //FRONT_END_GENERAL_CHECK(result, "CMO: Trackable Object Graph couldn't be read");
 
     for (const auto& node : tog.nodes()) {
         for (const auto& attr : node.attributes()) {
